Move response common functions to ResponseMaker trait
